# Core: Replace restricted_dumps with a RestrictedPickler implementation