## What is this fixing or adding?

Replaces `Utils.restricted_dumps` with a `Pickler` subclass that rejects objects that `RestrictedUnpickler` cannot unpickle.

It performs noticeably better when dumping multidata (.archipelago), but is about the same when dumping save data (.apsave).

I would guess it has a lower memory overhead, but I have not confirmed this.

## How was this tested?

I pickled and unpickled (using `Utils.restricted_dumps` and `Utils.restricted_loads`) multidata files (.archipelago) and save files (.apsave).
I also ran WebHost locally and both generated a game and hosted a large pre-generated game (sent locations in both cases in order to write to the save data and restarted the room successfully).

Attempting to pickle a non-allowed type raises an error like `_pickle.PicklingError: 'Al Kharid' (type '<enum 'RegionNames'>') is forbidden.`

Timings recorded using `timeit` on Python 3.12.9 on Windows 10, Ryzen 7 3800X
| Unpickled data and its uncompressed size before unpickling | trials | `pickle.dumps`/ms | `restricted_dumps` (current)/ms | `restricted_dumps` (PR)/ms |
| --- | --- | --- | --- | --- |
| multidata 11.5MB | 10 | 107 | 199 | 127 |
| multidata 60KB | 10000 | 0.255 | 0.540 | 0.264 |
| apsave 1.5MB | 10000 | 0.359 | 0.733 | 0.724 |
| apsave 36KB | 100000 | 0.00124 | 0.00559 | 0.00507 |

My guesses as to why the performance results are as they are:
Multidata (`.archipelago`) contains many builtin container types, `list`/`dict`/etc. filled primarily with `int` objects, all of which are pickled specially, so `RestrictedPickler` will never see them.

`.apsave` files are like 99% `NetUtils.NetworkItem` instances, so every instance will be intercepted by `RestrictedPickler`.

---

Throwing my hat into the ring with all the recent `restricted_dumps` discourse. I haven't really used `pickle` much, so there could be issues with this implementation that I am not aware of.

There is probably a way to set up both `RestrictedUnpickler` and `RestrictedPickler` such that they share a common input of what types/classes are allowed, but it sounds like it could be messy to avoid issues of cyclic dependencies, due to other modules needing to import `Utils`. If introducing a `RestrictedPickler` is a desired change, implementing a common input for both `RestrictedPickler` and `RestrictedUnpickler` could be something to look into.

I don't know of a reasonable way to compare memory usage between the old `restricted_dumps` and the `restricted_dumps` in this PR.