@duckboycool 2025-08-03
A basic memory test I did on my multidata sample with `tracemalloc.get_traced_memory` gave me ~4.2MB at the end and ~10.3MB peak when using regular `pickle.dumps`, ~4.4MB and ~22.8MB peak for current `restricted_dumps`, and ~4.2B with ~9.6MB peak with this implementation. So it seems that it does successfully help with the memory usage, and it might even be a bit better compared to the plain dumps.