@duckboycool 2025-08-03
Also I think this should be fine for usage here, but if the modules haven't already been imported prior to calling, then it seems to get into some weird import loop shenanigans and take way more time and memory.